refactor(e2e): relocate sequencer tests to single-node
Move the four e2e_sequencer tests (escape_hatch_vote_only, gov_proposal,
reload_keystore, slasher_config) into single-node/sequencer/ and convert
them onto SingleNodeTestContext via setupBlockProducer. Each preserves its
explicit opts (committee size, slot/epoch durations, proof-submission
window, governance round/quorum) and its PXE syncChainTip semantics
('proposed', matching the original raw setup() default). gov_proposal keeps
its .parallel suffix; the now-empty e2e_sequencer/ directory is removed.
